[TASK] Remove superfluous searchtools.js (#182)
This JavaScript file was used for the search in the Sphinx-based
rendering. Currently, we have no search available, so this file is
superfluous anyway. It is planned to implement a search with another
library independent of Sphinx.

Additionally, the autocompleter package has been removed as it was used
for autocompletion of search phrases. It can be re-added later - if
needed - when autocompletion for the new search will be implemented.